# c3-server-api/src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: Backend api utilized by the Chandler Care Center Customer registration
    system and self check-in kiosk.
  version: 3.0.0
  title: Chandler Care Center
  termsOfService: http://swagger.io/terms/
  contact:
    email: jacob.omaits@dcsg.com
    name: Dick's Sporting Goods
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /service/{service_id}:
    delete:
      tags:
        - Service
      summary: 'Deletes a given service.'
      parameters:
        - name: service_id
          in: path
          description: The service that should be deleted
          required: true
          schema:
            type: string
          example: 3fb6798c-c1cd-47d8-a123-49812f204cba
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/NotAllowed'
        '500':
          $ref: '#/components/responses/Exception'
...
  /services:
    get:
      tags:
        - Service
      summary: 'Gets a list of all services.'
      responses:
        '200':
          $ref: '#/components/responses/OK'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/NotAllowed'
        '500':
          $ref: '#/components/responses/Exception'
servers:
  - url: https://somedomain.com/v1
components:
...
  securitySchemes:
    c3_auth:
      type: oauth2
      description: Utilizes gmail federated login process to assign scopes.
      flows:
        implicit:
          authorizationUrl: http://google.com
          scopes:
            write:member: Create and update member data
            read:member: View member data
  responses:
    NotFound:
      description: The specified resource was not found.
...
    Exception:
      description: A system error has occurred and the request was not processed.
...
    BadRequest:
      description: The request could not be understood by the server due to malformed data.
    NotAllowed:
      description: The server has not implemented the requested HTTP method.
    OK:
      description: The request has processed successfully.
  schemas:
...
    Service:
      type: object
      properties:
        service_id:
          type: string
        name:
          type: string
      required:
        - name
...